# src/docs/swagger.yaml
basePath: /v1
definitions:
  github_com_TebanMT_smartGou_src_modules_reference_interfaces_lambdas_get_categories.ResponsePaginated:
    properties:
      categories:
        items:
          $ref: '#/definitions/main.Category'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_TebanMT_smartGou_src_modules_reference_interfaces_lambdas_get_meta_categories.ResponsePaginated:
    properties:
      limit:
        type: integer
      meta_categories:
        items:
          $ref: '#/definitions/main.MetaCategoryResponse'
        type: array
      offset:
        type: integer
      total:
        type: integer
    type: object
  main.Category:
    properties:
      color:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: string
      meta_category_id:
        type: string
      name_en:
        type: string
      name_es:
        type: string
    type: object
  main.ConfirmOtpRequest:
    properties:
      code:
        type: string
      user_id:
        type: string
    required:
    - code
    - user_id
    type: object
...
  main.MetaCategoryResponse:
    properties:
      color:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: string
      nameEn:
        type: string
      nameEs:
        type: string
    type: object
...
  main.SignUpByEmailRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  main.SignUpByEmailResponse:
    properties:
      user_id:
        type: string
    type: object
...
  utils.Response-any:
    properties:
      data: {}
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
  utils.Response-array_main_Category:
    properties:
      data:
        items:
          $ref: '#/definitions/main.Category'
        type: array
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
  utils.Response-array_main_MetaCategoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/main.MetaCategoryResponse'
        type: array
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
  utils.Response-github_com_TebanMT_smartGou_src_modules_reference_interfaces_lambdas_get_categories_ResponsePaginated:
    properties:
      data:
        $ref: '#/definitions/github_com_TebanMT_smartGou_src_modules_reference_interfaces_lambdas_get_categories.ResponsePaginated'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
  utils.Response-github_com_TebanMT_smartGou_src_modules_reference_interfaces_lambdas_get_meta_categories_ResponsePaginated:
    properties:
      data:
        $ref: '#/definitions/github_com_TebanMT_smartGou_src_modules_reference_interfaces_lambdas_get_meta_categories.ResponsePaginated'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  utils.Response-main_SignUpByEmailResponse:
    properties:
      data:
        $ref: '#/definitions/main.SignUpByEmailResponse'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
...
host: api.smartgo.com
info:
  contact: {}
  description: SmartGo API
  title: SmartGo API
  version: "1.0"
paths:
  /auth:
    patch:
      consumes:
      - application/json
      description: |-
        This endpoint verifies a OTP by email.
        Verify the code given a userID and a code. The code is the code that the user received in the email.
        Also, the email will be verified as true in provider (cognito) and database.
      parameters:
      - description: Request body
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.ConfirmOtpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response-any'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      summary: Login with email and password
      tags:
      - security
    post:
      consumes:
      - application/json
      description: |-
        This endpoint signs up a user by email and password.
        The user will be created in the database and inderity provider (cognito) if it doesn't exist.
        If the user exists, an error will be returned.
        A mail will be sent to the user to verify the email.
        If the user exists but the email is not verified, a new OTP will be sent to the user.
      parameters:
      - description: Request body
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.SignUpByEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-main_SignUpByEmailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      summary: Sign up by email and password
      tags:
      - security
...
  /references/categories:
    get:
      consumes:
      - application/json
      description: |-
        This endpoint gets the categories for a user. It can be filtered by meta category id or name like.
        If no filters are provided, all categories will be returned.
        If meta category id is provided, only categories related to that meta category will be returned.
        If name like is provided, only categories related to that name will be returned.
        If paged is true, you must provide valid limit and offset values.
        If paged is false, all categories will be returned, ignoring limit and offset values.
        If order by is provided, the order of the categories will be based on the order by field and order dir field will be required.
        If order by is not provided, the categories will be returned in the order of the database.
      parameters:
      - description: Meta Category ID
        in: query
        name: meta_category_id
        type: string
      - description: Name Like
        in: query
        name: name_like
        type: string
      - description: Order By
        in: query
        name: order_by
        type: string
      - description: Order Dir
        in: query
        name: order_dir
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Paged
        in: query
        name: paged
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-github_com_TebanMT_smartGou_src_modules_reference_interfaces_lambdas_get_categories_ResponsePaginated'
        "206":
          description: The code 206 just means the response has not been paginated
            but the response will send with 200 code
          schema:
            $ref: '#/definitions/utils.Response-array_main_Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      security:
      - BearerAuth: []
      summary: Get Categories
      tags:
      - reference
  /references/meta-categories:
    get:
      consumes:
      - application/json
      description: |-
        This endpoint gets the meta categories catalog.
        If no filters are provided, all meta categories will be returned.
        If name like is provided, only meta categories related to that name will be returned.
        If paged is true, you must provide valid limit and offset values.
        If paged is false, all meta categories will be returned, ignoring limit and offset values.
        If order by is provided, the order of the meta categories will be based on the order by field and order dir field will be required.
        If order by is not provided, the meta categories will be returned in the order of the database.
      parameters:
      - description: Name Like
        in: query
        name: name_like
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Paged
        in: query
        name: paged
        type: boolean
      - description: Order By
        in: query
        name: order_by
        type: string
      - description: Order Dir
        in: query
        name: order_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-github_com_TebanMT_smartGou_src_modules_reference_interfaces_lambdas_get_meta_categories_ResponsePaginated'
        "206":
          description: The code 206 just means the response has not been paginated
            but the response will send with 200 code
          schema:
            $ref: '#/definitions/utils.Response-array_main_MetaCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      security:
      - BearerAuth: []
      summary: Get meta categories
      tags:
      - reference
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"